dom/tests/mochitest/bugs/test_bug809290.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=809290
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 809290</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=809290">Mozilla Bug 809290</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/*
 * Test for Bug 809290.
 *
 * This test used to check that the referrer was the entry settings object.
 * Due to spec changes, we now need to check that the referrer is the incumbent
 * settings object.
 */
SimpleTest.waitForExplicitFinish();

var gNotifyCount = 0;
var base = window.location.href.replace(/test_bug.*/, '');
function notifyReferrer(referrer) {
  ++gNotifyCount;
  if (gNotifyCount == 1) {
    is(referrer, base + 'file_bug809290_b1.html', "Referrer should come from the incumbent script settings object (1)");
    document.getElementById('ifr').setAttribute('src', 'file_bug809290_b2.html');
  } else {
    is(gNotifyCount, 2, "notify count");
    is(referrer, base + 'file_bug809290_b2.html', "Referrer should come from the the incumbent script settings object (2)");
    SimpleTest.finish();
  }
}

function go() {
  var ifr = document.getElementById('ifr');
  ifr.onload = null;
  ifr.contentWindow.innerLoad();
}

</script>
</pre>
<iframe id="ifr" src="file_bug809290_b1.html" onload="go();"></iframe>
</body>
</html>
